move cluster.rs to new cluster mod
