+++ scylla/src/cluster/cluster.rs
@@ -6,6 +6,7 @@ use crate::routing::{Shard, Token};
 use crate::session::TABLET_CHANNEL_SIZE;
 use crate::transport::errors::{BadQuery, NewSessionError, QueryError};
 use crate::transport::host_filter::HostFilter;
+use crate::transport::locator::ReplicaLocator;
 use crate::transport::node::Node;
 use crate::transport::partitioner::PartitionerName;
 
@@ -22,14 +23,10 @@ use std::time::Duration;
 use tracing::{debug, warn};
 use uuid::Uuid;
 
-use super::locator::tablets::{RawTablet, Tablet, TabletsInfo};
-use super::node::{InternalKnownNode, NodeAddr};
-
-use super::NodeRef;
 use crate::cluster::metadata::{Keyspace, Metadata, MetadataReader, Strategy};
-
-use super::locator::ReplicaLocator;
-use super::partitioner::calculate_token_for_partition_key;
+use crate::transport::locator::tablets::{RawTablet, Tablet, TabletsInfo};
+use crate::transport::node::{InternalKnownNode, NodeAddr, NodeRef};
+use crate::transport::partitioner::calculate_token_for_partition_key;
 
 /// Cluster manages up to date information and connections to database nodes.
 /// All data can be accessed by cloning Arc<ClusterData> in the `data` field
