+++ scylla/src/cluster/mod.rs
@@ -1,1 +1,6 @@
+#[allow(clippy::module_inception)]
+mod cluster;
 pub mod metadata;
+
+pub use cluster::ClusterData;
+pub(crate) use cluster::{use_keyspace_result, Cluster, ClusterNeatDebug};
