+++ scylla/src/connection/connection_pool.rs
@@ -1106,7 +1106,7 @@ impl PoolRefiller {
             .await
             .map_err(|_| QueryError::TimeoutError)?;
 
-            crate::transport::cluster::use_keyspace_result(use_keyspace_results.into_iter())
+            crate::cluster::use_keyspace_result(use_keyspace_results.into_iter())
         };
 
         tokio::task::spawn(async move {
