+++ scylla/src/session/session.rs
@@ -36,12 +36,12 @@ use tokio::time::timeout;
 use tracing::{debug, error, trace, trace_span, Instrument};
 use uuid::Uuid;
 
-use crate::connection::Compression;
+use crate::cluster::{Cluster, ClusterData, ClusterNeatDebug};
 #[cfg(feature = "ssl")]
 use crate::connection::SslConfig;
-use crate::connection::{AddressTranslator, QueryResponse};
+use crate::connection::{AddressTranslator, NonErrorQueryResponse, SelfIdentity};
+use crate::connection::{Compression, QueryResponse};
 use crate::connection::{Connection, ConnectionConfig, VerifiedKeyspaceName};
-use crate::connection::{NonErrorQueryResponse, SelfIdentity};
 use crate::connection::{PoolConfig, PoolSize};
 use crate::frame::response::result;
 use crate::prepared_statement::PreparedStatement;
@@ -49,7 +49,6 @@ use crate::query::Query;
 use crate::routing::{Shard, Token};
 use crate::statement::{Consistency, PageSize, PagingState, PagingStateResponse};
 use crate::tracing::TracingInfo;
-use crate::transport::cluster::{Cluster, ClusterData, ClusterNeatDebug};
 use crate::transport::errors::TracingProtocolError;
 use crate::transport::execution_profile::{
     ExecutionProfile, ExecutionProfileHandle, ExecutionProfileInner,
