+++ scylla/src/transport/iterator.rs
@@ -9,17 +9,20 @@ use std::task::{Context, Poll};
 
 use futures::Stream;
 use scylla_cql::frame::frame_errors::ResultMetadataAndRowsCountParseError;
+use scylla_cql::frame::request::query::PagingState;
 use scylla_cql::frame::response::result::RawMetadataAndRawRows;
 use scylla_cql::frame::response::NonErrorResponse;
+use scylla_cql::frame::types::SerialConsistency;
 use scylla_cql::types::deserialize::result::RawRowLendingIterator;
 use scylla_cql::types::deserialize::row::{ColumnIterator, DeserializeRow};
 use scylla_cql::types::deserialize::{DeserializationError, TypeCheckError};
 use scylla_cql::types::serialize::row::SerializedValues;
+use scylla_cql::Consistency;
 use std::result::Result;
 use thiserror::Error;
 use tokio::sync::mpsc;
 
-use super::errors::UserRequestError;
+use super::errors::{QueryError, UserRequestError};
 use super::execution_profile::ExecutionProfileInner;
 use super::query_result::ColumnSpecs;
 #[allow(deprecated)]
@@ -27,6 +30,7 @@ use crate::cql_to_rust::{FromRow, FromRowError};
 use crate::deserialize::DeserializeOwnedRow;
 use crate::session::RequestSpan;
 
+use crate::cluster::ClusterData;
 use crate::connection::{Connection, NonErrorQueryResponse, QueryResponse};
 use crate::frame::response::{
     result,
@@ -34,9 +38,7 @@ use crate::frame::response::{
 };
 use crate::history::{self, HistoryListener};
 use crate::statement::{prepared_statement::PreparedStatement, query::Query};
-use crate::statement::{Consistency, PagingState, SerialConsistency};
-use crate::transport::cluster::ClusterData;
-use crate::transport::errors::{ProtocolError, QueryError};
+use crate::transport::errors::ProtocolError;
 use crate::transport::load_balancing::{self, RoutingInfo};
 use crate::transport::metrics::Metrics;
 use crate::transport::retry_policy::{QueryInfo, RetryDecision, RetrySession};
