+++ scylla/tests/integration/consistency.rs
@@ -377,7 +377,7 @@ impl LoadBalancingPolicy for RoutingInfoReportingWrapper {
     fn pick<'a>(
         &'a self,
         query: &'a RoutingInfo,
-        cluster: &'a scylla::transport::ClusterData,
+        cluster: &'a scylla::cluster::ClusterData,
     ) -> Option<(NodeRef<'a>, Option<Shard>)> {
         self.routing_info_tx
             .send(OwnedRoutingInfo::from(query.clone()))
@@ -388,7 +388,7 @@ impl LoadBalancingPolicy for RoutingInfoReportingWrapper {
     fn fallback<'a>(
         &'a self,
         query: &'a RoutingInfo,
-        cluster: &'a scylla::transport::ClusterData,
+        cluster: &'a scylla::cluster::ClusterData,
     ) -> scylla::load_balancing::FallbackPlan<'a> {
         self.routing_info_tx
             .send(OwnedRoutingInfo::from(query.clone()))
