+++ scylla/tests/integration/execution_profiles.rs
@@ -10,10 +10,10 @@ use scylla::routing::Shard;
 use scylla::statement::SerialConsistency;
 use scylla::transport::NodeRef;
 use scylla::{
+    cluster::ClusterData,
     load_balancing::{LoadBalancingPolicy, RoutingInfo},
     retry_policy::{RetryPolicy, RetrySession},
     speculative_execution::SpeculativeExecutionPolicy,
-    transport::ClusterData,
     ExecutionProfile, SessionBuilder,
 };
 use scylla_cql::Consistency;
